requirements.txt
requests
tqdm
pillow==10.2.0
transformers==4.41.2
transformers-stream-generator==0.0.5
einops==0.8.0
matplotlib
accelerate==0.31.0
chardet==5.2.0
flash-attn
tiktoken==0.7.0
xformers
openai-clip
apex
bert-score
bitsandbytes
black
brotli
cachelib
cloudpickle
decord
fairseq
faiss-gpu
future
fvcore
gmpy2
httptools
humanize
ipykernel
mkl-fft
mkl-random
modelscope
opencv-python
openpyxl
pydot
pysocks
rapidfuzz
triton
unicodedata2
uvloop
watchfiles
xformers
zhconv
nltk
